Webb/Back-End/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==1.13.1
torchvision==0.14.1
numpy==1.23.4
opencv-python==4.7.0.72
pillow==9.3.0
albumentations==1.3.0
tqdm==4.64.1
matplotlib==3.6.2 